fixed 64-bit memory pointer warnings
Cast pointers to integers only of the same size as the integer type big enough to hold a pointer address.

+++ Source/Project64-core/N64System/Mips/MemoryVirtualMem.cpp
@@ -110,7 +110,7 @@ void CMipsMemoryVM::Reset(bool /*EraseMemory*/)
 
 void CMipsMemoryVM::ReserveMemory()
 {
-    m_Reserve1 = (uint8_t *)AllocateAddressSpace(0x20000000, (void *)g_Settings->LoadDword(Setting_FixedRdramAddress));
+    m_Reserve1 = (uint8_t *)AllocateAddressSpace(0x20000000, (void *)(uintptr_t)g_Settings->LoadDword(Setting_FixedRdramAddress));
     m_Reserve2 = (uint8_t *)AllocateAddressSpace(0x04002000);
 }
