+++ Source/Project64-core/N64System/Mips/Sram.cpp
@@ -62,7 +62,7 @@ void CSram::DmaFromSram(uint8_t * dest, int32_t StartOffset, int32_t len)
     // Fix Dezaemon 3D saves
     StartOffset = ((StartOffset >> 3) & 0xFFFF8000) | (StartOffset & 0x7FFF);
 
-    if (((StartOffset & 3) == 0) && ((((uint32_t)dest) & 3) == 0))
+    if (((StartOffset & 3) == 0) && ((((uintptr_t)dest) & 3) == 0))
     {
         m_File.Seek(StartOffset, CFile::begin);
         m_File.Read(dest, len);
@@ -72,7 +72,7 @@ void CSram::DmaFromSram(uint8_t * dest, int32_t StartOffset, int32_t len)
         for (i = 0; i < len; i++)
         {
             m_File.Seek((StartOffset + i) ^ 3, CFile::begin);
-            m_File.Read((uint8_t*)(((uint32_t)dest + i) ^ 3), 1);
+            m_File.Read((uint8_t*)(((uintptr_t)dest + i) ^ 3), 1);
         }
     }
 }
@@ -97,7 +97,7 @@ void CSram::DmaToSram(uint8_t * Source, int32_t StartOffset, int32_t len)
     // Fix Dezaemon 3D saves
     StartOffset = ((StartOffset >> 3) & 0xFFFF8000) | (StartOffset & 0x7FFF);
 
-    if (((StartOffset & 3) == 0) && ((((uint32_t)Source) & 3) == 0) && NULL != NULL)
+    if (((StartOffset & 3) == 0) && ((((uintptr_t)Source) & 3) == 0) && NULL != NULL)
     {
         m_File.Seek(StartOffset, CFile::begin);
         m_File.Write(Source, len);
@@ -107,7 +107,7 @@ void CSram::DmaToSram(uint8_t * Source, int32_t StartOffset, int32_t len)
         for (i = 0; i < len; i++)
         {
             m_File.Seek((StartOffset + i) ^ 3, CFile::begin);
-            m_File.Write((uint8_t*)(((uint32_t)Source + i) ^ 3), 1);
+            m_File.Write((uint8_t*)(((uintptr_t)Source + i) ^ 3), 1);
         }
     }
 }
